refactor: extract SimulationState dataclasses to simulation/state.py (Architect Rec 3)
Move InteractionRecord, CommunityOfInquiryState, and SimulationState out of
engine.py into a new state.py module. Update all 14 import sites across
data_output and theories packages to import directly from simulation.state.
