+++ synthed/simulation/theories/bean_metzner.py
@@ -7,7 +7,7 @@ import numpy as np
 
 if TYPE_CHECKING:
     from ...agents.persona import StudentPersona
-    from ..engine import SimulationState
+    from ..state import SimulationState
 
 
 class BeanMetznerPressure:
