+++ synthed/simulation/theories/moore_td.py
@@ -7,7 +7,7 @@ import numpy as np
 
 if TYPE_CHECKING:
     from ...agents.persona import StudentPersona
-    from ..engine import SimulationState
+    from ..state import SimulationState
     from ..environment import Course, ODLEnvironment
