+++ synthed/simulation/theories/positive_events.py
@@ -11,7 +11,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ..engine import SimulationState
+    from ..state import SimulationState
     from ...agents.persona import StudentPersona
