+++ synthed/simulation/theories/protocol.py
@@ -22,7 +22,7 @@ import numpy as np
 
 if TYPE_CHECKING:
     from ...agents.persona import StudentPersona
-    from ..engine import InteractionRecord, SimulationState
+    from ..state import InteractionRecord, SimulationState
     from ..engine_config import EngineConfig
     from ..environment import Course, ODLEnvironment
     from ..institutional import InstitutionalConfig
